+++ app/ide-desktop/lib/dashboard/src/authentication/src/http.tsx
@@ -35,25 +35,26 @@ function blobToBase64(blob: Blob) {
 
 /** An HTTP client that can be used to create and send HTTP requests asynchronously. */
 export class Client {
+    /** Create a new HTTP client with the specified headers to be sent on every request. */
     constructor(
         /** A map of default headers that are included in every HTTP request sent by this client.
          *
-         * This is useful for setting headers that are required for every request, like authentication
-         * tokens. */
-        public defaultHeaders?: Headers
+         * This is useful for setting headers that are required for every request, like
+         * authentication tokens. */
+        public defaultHeaders: Headers
     ) {}
 
-    /** Sends an HTTP GET request to the specified URL. */
+    /** Send an HTTP GET request to the specified URL. */
     get<T = void>(url: string) {
         return this.request<T>(HttpMethod.get, url)
     }
 
-    /** Sends a JSON HTTP POST request to the specified URL. */
+    /** Send a JSON HTTP POST request to the specified URL. */
     post<T = void>(url: string, payload: object) {
         return this.request<T>(HttpMethod.post, url, JSON.stringify(payload), 'application/json')
     }
 
-    /** Sends a base64-encoded binary HTTP POST request to the specified URL. */
+    /** Send a base64-encoded binary HTTP POST request to the specified URL. */
     async postBase64<T = void>(url: string, payload: Blob) {
         return await this.request<T>(
             HttpMethod.post,
@@ -63,32 +64,34 @@ export class Client {
         )
     }
 
-    /** Sends a JSON HTTP PUT request to the specified URL. */
+    /** Send a JSON HTTP PUT request to the specified URL. */
     put<T = void>(url: string, payload: object) {
         return this.request<T>(HttpMethod.put, url, JSON.stringify(payload), 'application/json')
     }
 
-    /** Sends an HTTP DELETE request to the specified URL. */
+    /** Send an HTTP DELETE request to the specified URL. */
     delete<T = void>(url: string) {
         return this.request<T>(HttpMethod.delete, url)
     }
 
-    /** Executes an HTTP request to the specified URL, with the given HTTP method. */
+    /** Execute an HTTP request to the specified URL, with the given HTTP method. */
     private request<T = void>(
         method: HttpMethod,
         url: string,
         payload?: string,
         mimetype?: string
     ) {
-        const defaultHeaders = this.defaultHeaders ?? []
-        const headers = new Headers(defaultHeaders)
+        const headers = new Headers(this.defaultHeaders)
         if (payload) {
             const contentType = mimetype ?? 'application/json'
             headers.set('Content-Type', contentType)
         }
+
+        /** A {@link Response} with a properly typed return type for `response.json()`. */
         interface ResponseWithTypedJson<U> extends Response {
             json: () => Promise<U>
         }
+
         // This is an UNSAFE type assertion, however this is a HTTP client
         // and should only be used to query APIs with known response types.
         // eslint-disable-next-line no-restricted-syntax